--- lib/config/constants.dart
@@ -1,6 +1,0 @@
-const double kDefaultPadding = 16;
-const double kDefaultGapSize = 4;
-
-const double kDefaultRadius = 12;
-
-const Duration k300MS = Duration(milliseconds: 300);
