--- lib/model/difficulty.dart
@@ -1,1 +1,0 @@
-enum Difficulty { easy, medium, hard }
