--- lib/util/only_web/only_web.dart
@@ -1,2 +1,0 @@
-export 'only_web_interface.dart'
-    if (dart.library.html) 'only_web_interface_achieve.dart';
