--- lib/util/only_web/only_web_interface.dart
@@ -1,1 +1,0 @@
-void launchUrlString(String _) {}
