--- lib/util/only_web/only_web_interface_achieve.dart
@@ -1,5 +1,0 @@
-import 'dart:html';
-
-void launchUrlString(String urlString) {
-  window.open(urlString, 'New Tab');
-}
